[Tester\TesterBundle] Refactor PHPUnit extension and improve documentation
