--- packages/tester-bundle/PHPUnit/Extension/SetUpAutowire/AutowiredInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire;
-
-interface AutowiredInterface
-{
-}
