+++ packages/tester-bundle/extension.neon
@@ -1,7 +1,7 @@
 services:
-    draw.tester_bundle.autowire_client_read_write_properties_extension:
-        class: Draw\Bundle\TesterBundle\PHPStan\Rules\Properties\AutowireReadWritePropertiesExtension
+    draw.component.tester.autowire_client_read_write_properties_extension:
+        class: Draw\Component\Tester\PHPStan\Rules\Properties\AutowireReadWritePropertiesExtension
         arguments:
-            - 'Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireInterface'
+            - 'Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowireInterface'
         tags:
             - phpstan.properties.readWriteExtension
