+++ packages/tester/PHPUnit/Extension/SetUpAutowire/AutowiredInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire;
+
+interface AutowiredInterface
+{
+}
