+++ tests/Controller/Api/TestDocumentControllerTest.php
@@ -5,8 +5,8 @@ namespace App\Tests\Controller\Api;
 use App\Message\NewTestDocumentMessage;
 use Draw\Bundle\TesterBundle\Messenger\MessengerTesterTrait;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireClient;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Bundle\TesterBundle\WebTestCase;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 
 class TestDocumentControllerTest extends WebTestCase implements AutowiredInterface
