+++ tests/Controller/OpenApiControllerTest.php
@@ -3,8 +3,8 @@
 namespace App\Tests\Controller;
 
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireClient;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Bundle\TesterBundle\WebTestCase;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 
 class OpenApiControllerTest extends WebTestCase implements AutowiredInterface
