+++ tests/OpenApi/EventListener/ResponseApiExceptionListenerTest.php
@@ -3,10 +3,10 @@
 namespace App\Tests\OpenApi\EventListener;
 
 use Draw\Bundle\TesterBundle\EventDispatcher\EventListenerTestTrait;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireService;
 use Draw\Component\OpenApi\Event\PreDumpRootSchemaEvent;
 use Draw\Component\OpenApi\EventListener\ResponseApiExceptionListener;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
 class ResponseApiExceptionListenerTest extends KernelTestCase implements AutowiredInterface
