+++ tests/SonataExtraBundle/EventListener/PreObjectDeleteBatchEventEventListenerTest.php
@@ -6,8 +6,8 @@ use App\Entity\User;
 use App\Security\Voter\CannotSelfVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Draw\Bundle\SonataExtraBundle\EventListener\PreObjectDeleteBatchEventEventListener;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireService;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Sonata\DoctrineORMAdminBundle\Event\PreObjectDeleteBatchEvent;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
