+++ tests/SonataIntegrationBundle/Messenger/Action/RetryFailedMessageActionTest.php
@@ -8,9 +8,9 @@ use App\Entity\MessengerMessage;
 use App\Message\FailedMessage;
 use App\Tests\SonataIntegrationBundle\WebTestCaseTrait;
 use Draw\Bundle\TesterBundle\Messenger\MessengerTesterTrait;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireService;
 use Draw\Component\Messenger\Message\RetryFailedMessageMessage;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Stamp\TransportNamesStamp;
