+++ tests/TesterBundle/PHPUnit/Extension/SetUpAutoWire/SetUpAutowireExtensionTest.php
@@ -6,12 +6,12 @@ use App\EntityMigration\UserSetCommentNullMigration;
 use Doctrine\Persistence\ManagerRegistry;
 use Draw\Bundle\TesterBundle\Messenger\TransportTester;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireClient;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireMock;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireMockProperty;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireParameter;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireService;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireTransportTester;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowireMock;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowireMockProperty;
 use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
