+++ tests/Validator/Constraints/ValueIsNotUsedValidatorTest.php
@@ -4,8 +4,8 @@ namespace App\Tests\Validator\Constraints;
 
 use App\Entity\Tag;
 use App\Entity\User;
-use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Bundle\TesterBundle\PHPUnit\Extension\SetUpAutowire\AutowireService;
+use Draw\Component\Tester\PHPUnit\Extension\SetUpAutowire\AutowiredInterface;
 use Draw\Component\Validator\Constraints\ValueIsNotUsed;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
